fix: stop three languages fusing a multiply and an add
An estimator here is defined as a fixed sequence of binary64 operations. A fused
multiply-add performs two of them with one rounding, which is a legitimate IEEE 754
value and still differs in the last bit from the six languages that cannot fuse.
Three ports could, each for a different reason, and two of them already did.

Rust diverged everywhere. Clippy's `suboptimal_flops` advice had rewritten
`min + (max - min) * u` as `(max - min).mul_add(u, min)`, which lowers to
llvm.fma.f64 and rounds once where the specification rounds twice.
`Rng::from_seed(1729).uniform_f64_range(-50.0, 50.0)` returned
-10.569652967034637 against the -10.569652967034635 in the shipped fixture and in
the other six; six of that fixture's twenty values differed, and a 1e-12 comparison
called it a pass for months. Every mul_add is reverted, not only the three under a
bitwise contract.

Go diverged on the targets CI does not run. The specification permits the fusion
and gc performs it on arm64, ppc64le, s390x, riscv64 and loong64, never on amd64.
Measured under qemu-aarch64, `pairwiseMargin` returned a different integer on 8268
of 79797 (n, m) pairs at the minimum achievable misrate: the comparison there
reduces to an exact equality decided purely by rounding, and the exponential turns
one ulp of a logarithm into a different index. Every product feeding an add is now
wrapped in an explicit conversion, including the ones scaling by a power of two,
where fusion is provably exact - an exemption every future reader would have to
re-derive is worth less than a conversion.

R's C kernels had inherited the fusion from the toolchain. The C standard defaults
to `FP_CONTRACT ON`, gcc to `-ffp-contract=fast`, and clang has defaulted to `on`
since version 14; with no Makevars the kernels took that verbatim from Makeconf.
Recompiling all four with `-march=x86-64-v3`, which has FMA3 and stands in for
aarch64, produced twelve fused instructions; `-ffp-contract=off` produces none.

Acked-by: Andrey Akinshin <andrey.akinshin@gmail.com>

+++ go/additive.go
@@ -35,7 +35,12 @@ func (a *Additive) Sample(rng *Rng) float64 {
 	// Use the first of the two Box-Muller outputs
 	z := r * math.Cos(theta)
 
-	return a.Mean + z*a.StdDev
+	// float64 conversion pins the intermediate rounding; see uniformFloat64Range
+	// in xoshiro256.go. It does not make this draw portable on its own: math.Log
+	// and math.Cos are themselves polynomial evaluations in Go, and gc fuses them
+	// on FMA-capable targets, so a residual divergence remains that no conversion
+	// in this file can remove.
+	return a.Mean + float64(z*a.StdDev)
 }
 
 // Samples generates multiple samples from the additive distribution.
